make BAM lookup by name work when reads in a pair actually have different names
This assumes that names for reads in pair are identical with the exception of
a few characters and that the position of the different characters is the same
for all reads. The first read pair in a FASTQ file is used to determine the
positions containing the differences. These characters are then removed from
the read name for BAM indexing and lookup.